refactor(ui): internalize chat update scheduling
